requirements.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
PyMuPDF
python-docx
python-dotenv
tiktoken
